mistral/requirements.txt
Flask==2.3.3
mistralai==0.1.4
python-dotenv==1.0.0